Code cleanup. Moved all interfaces to the "helper" folder under the mapper.
