+++ src/gd-sprest.ts
@@ -69,7 +69,7 @@ export {
  * SharePoint REST Library
  */
 export const $REST: Types.IREST = {
-    __ver: 2.66,
+    __ver: 2.67,
     ContextInfo,
     DefaultRequestToHostFl: false,
     Helper,
