+++ src/helper/index.ts
@@ -1,16 +1,16 @@
 import { Types } from "../mapper";
 import { IBase } from "../utils";
 import { App } from "./app";
-import { Dependencies, IDependencies } from "./dependencies";
+import { Dependencies } from "./dependencies";
 import { FieldSchemaXML } from "./field";
-import { JSLinkHelper, IHelperJSLink } from "./jslink";
+import { JSLink } from "./jslink";
 import { ListForm } from "./listForm";
 import { ListFormField } from "./listFormField";
-import { Loader, ILoader } from "./loader";
+import { Loader } from "./loader";
 import { parse } from "./parse";
-import { SPConfig, ISPConfig } from "./spCfg";
-import { HelperTypes, IHelperTypes } from "./types";
-import { WebPart, IWebPart } from "./webpart";
+import { SPConfig } from "./spCfg";
+import { HelperTypes } from "./types";
+import { WebPart } from "./webpart";
 
 /**
  * Helper
@@ -24,7 +24,7 @@ export interface IHelper {
     /**
      * Dependencies
      */
-    Dependencies: IDependencies,
+    Dependencies: Types.Helper.Dependencies.IDependencies,
 
     /**
      * Field Schema XML
@@ -34,7 +34,7 @@ export interface IHelper {
     /**
      * JSLink helper methods
      */
-    JSLink: IHelperJSLink,
+    JSLink: Types.Helper.JSLink.IJSLink,
 
     /**
      * List Form
@@ -49,7 +49,7 @@ export interface IHelper {
     /**
      * Loader
      */
-    Loader: ILoader,
+    Loader: Types.Helper.Loader.ILoader,
 
     /**
      * Method to parse a json string and convert to a base object.
@@ -59,17 +59,17 @@ export interface IHelper {
     /**
      * Web helper methods
      */
-    SPConfig: ISPConfig,
+    SPConfig: Types.Helper.SPConfig.ISPConfig,
 
     /**
      * Helper Types
      */
-    Types: IHelperTypes
+    Types: Types.Helper.IHelperTypes;
 
     /**
      * WebPart
      */
-    WebPart: IWebPart;
+    WebPart: Types.Helper.WebPart.IWebPart;
 };
 
 /**
@@ -77,14 +77,14 @@ export interface IHelper {
  */
 export const Helper: IHelper = {
     App,
-    Dependencies: Dependencies as any,
+    Dependencies,
     FieldSchemaXML,
-    JSLink: JSLinkHelper,
+    JSLink,
     ListForm,
     ListFormField,
-    Loader: Loader,
+    Loader,
     parse: parse as any,
-    SPConfig: SPConfig as any,
+    SPConfig,
     Types: HelperTypes,
     WebPart
 };
