+++ src/helper/loader.ts
@@ -1,22 +1,10 @@
 import { ContextInfo } from "../lib";
+import { Types } from "../mapper";
 
 /**
  * Loader
  */
-export interface ILoader {
-    /**
-     * Waits for the SharePoint core libraries to be loaded.
-     * @param callback - The callback function.
-     * @param timeout - The max time (ms) to wait for the libraries to be loaded.
-     * @param loadLibraries - Flag to load the core libraries manually.
-     */
-    waitForSPLibs(callback: any, timeout?: number, loadLibraries?: boolean);
-}
-
-/**
- * Loader
- */
-export const Loader = {
+export const Loader: Types.Helper.Loader.ILoader = {
     // Flag to determine if the SharePoint core libraries have been loaded
     loaded: false,
