+++ src/helper/types.ts
@@ -1,3 +1,5 @@
+import { Types } from "../mapper";
+
 /**
  * SharePoint Configuration Field Types
  */
@@ -31,37 +33,7 @@ const SPCfgType = {
 /**
  * Helper Types
  */
-export interface IHelperTypes {
-    /** The field types */
-    SPCfgFieldType: {
-        Boolean: number;
-        Calculated: number;
-        Choice: number;
-        Date: number;
-        Lookup: number;
-        MMS: number;
-        Note: number;
-        Number: number;
-        Text: number;
-        Url: number;
-        User: number;
-    };
-
-    /** The configuration types */
-    SPCfgType: {
-        Fields: number;
-        ContentTypes: number;
-        Lists: number;
-        SiteUserCustomActions: number;
-        WebParts: number;
-        WebUserCustomActions: number;
-    };
-}
-
-/**
- * Helper Types
- */
-export const HelperTypes: IHelperTypes = {
+export const HelperTypes: Types.Helper.IHelperTypes = {
     SPCfgFieldType,
     SPCfgType
 }
