+++ src/mapper/helper/index.ts
@@ -1,12 +1,19 @@
 import * as App from "./app";
+import * as Dependencies from "./dependencies";
 import * as Field from "./field";
+import * as JSLink from "./jslink";
 import * as ListForm from "./listForm";
+import * as Loader from "./loader";
 import * as SPConfig from "./spCfg";
 import * as WebPart from "./webpart";
+export * from "./types";
 export {
     App,
+    Dependencies,
     Field,
+    JSLink,
     ListForm,
+    Loader,
     SPConfig,
     WebPart
 };
